Devex 837: rename AddDocketEntry to AddPaperFiling
