+++ web-client/integration-tests/journey/docketClerkEditsDocketEntryNonstandardA.js
@@ -35,7 +35,7 @@ export const docketClerkEditsDocketEntryNonstandardA = test => {
       docketNumber: test.docketNumber,
     });
 
-    expect(test.getState('currentPage')).toEqual('AddDocketEntry');
+    expect(test.getState('currentPage')).toEqual('AddPaperFiling');
     expect(test.getState('docketEntryId')).toEqual(docketEntryId);
 
     expect(test.getState('form')).toMatchObject({
