+++ web-client/integration-tests/journey/docketClerkEditsDocketEntryNonstandardC.js
@@ -35,7 +35,7 @@ export const docketClerkEditsDocketEntryNonstandardC = test => {
       docketNumber: test.docketNumber,
     });
 
-    expect(test.getState('currentPage')).toEqual('AddDocketEntry');
+    expect(test.getState('currentPage')).toEqual('AddPaperFiling');
     expect(test.getState('docketEntryId')).toEqual(docketEntryId);
 
     await test.runSequence('updateDocketEntryFormValueSequence', {
