+++ web-client/integration-tests/journey/docketClerkEditsDocketEntryNonstandardH.js
@@ -37,7 +37,7 @@ export const docketClerkEditsDocketEntryNonstandardH = test => {
       docketNumber: test.docketNumber,
     });
 
-    expect(test.getState('currentPage')).toEqual('AddDocketEntry');
+    expect(test.getState('currentPage')).toEqual('AddPaperFiling');
     expect(test.getState('docketEntryId')).toEqual(docketEntryId);
 
     await test.runSequence('updateDocketEntryFormValueSequence', {
