+++ web-client/integration-tests/journey/docketClerkEditsDocketEntryStandard.js
@@ -30,7 +30,7 @@ export const docketClerkEditsDocketEntryStandard = test => {
       docketNumber: test.docketNumber,
     });
 
-    expect(test.getState('currentPage')).toEqual('AddDocketEntry');
+    expect(test.getState('currentPage')).toEqual('AddPaperFiling');
     expect(test.getState('docketEntryId')).toEqual(docketEntryId);
 
     await test.runSequence('updateDocketEntryFormValueSequence', {
