+++ web-client/integration-tests/journey/docketClerkViewsEditDocketRecord.js
@@ -9,6 +9,6 @@ export const docketClerkViewsEditDocketRecord = test => {
       docketNumber: test.docketNumber,
     });
 
-    expect(test.getState('currentPage')).toEqual('AddDocketEntry');
+    expect(test.getState('currentPage')).toEqual('AddPaperFiling');
   });
 };
